Move ListItemData to applist package
