+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/applist/ListItemData.kt
@@ -1,8 +1,9 @@
-package net.mullvad.mullvadvpn.model
+package net.mullvad.mullvadvpn.applist
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.IntDef
 import androidx.annotation.StringRes
+import net.mullvad.mullvadvpn.model.WidgetState
 
 data class ListItemData
 private constructor(
