+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/applist/ViewIntent.kt
@@ -1,7 +1,5 @@
 package net.mullvad.mullvadvpn.applist
 
-import net.mullvad.mullvadvpn.model.ListItemData
-
 sealed class ViewIntent {
     // In future we will have search intent
     data class ChangeApplicationGroup(val item: ListItemData) : ViewIntent()
