+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/ui/ListItemListener.kt
@@ -1,6 +1,6 @@
 package net.mullvad.mullvadvpn.ui
 
-import net.mullvad.mullvadvpn.model.ListItemData
+import net.mullvad.mullvadvpn.applist.ListItemData
 
 interface ListItemListener {
     fun onItemAction(item: ListItemData)
